refactor: Componentize QTable for reuse
